Implement Equipment Addition
-implement equipment purchase options dialog
-implement max gold input
-implement item purchasing
-implement spent currency display
-change cost integers to doubles
-implement viewModel for user implemented inputs and displays
-change how character intent is applied

+++ app/src/main/java/com/example/animabuilder/character_creation/equipment/general_goods/GeneralCategory.kt
@@ -19,7 +19,7 @@ open class GeneralCategory(val qualityInput: List<QualityModifier>?): Serializab
      * @return the exact item held in inventory, a copy of that item with a different cost, or
      * an empty flag indicating no item matching the inputted name
      */
-    fun findEquipment(name: String, cost: Int): GeneralEquipment?{
+    fun findEquipment(name: String, cost: Double): GeneralEquipment?{
         //look at each available item
         itemsAvailable.forEach{
             //item name found
