+++ app/src/main/java/com/example/animabuilder/character_creation/equipment/general_goods/GeneralEquipment.kt
@@ -5,7 +5,7 @@ import java.io.Serializable
 
 class GeneralEquipment(
     val name: String,
-    val baseCost: Int,
+    val baseCost: Double,
     val coinType: CoinType,
     val weight: Double?,
     val availability: Availability
